prepare for release - next version - test util moved to util directory
