+++ .env
@@ -1,1 +1,1 @@
-DICT_VERSION=0.0.18
+DICT_VERSION=0.0.19
