+++ calendar/repository_test.go
@@ -1,8 +1,8 @@
 package calendar
 
 import (
-	"dictionaries-service/test"
 	"dictionaries-service/util"
+	test2 "dictionaries-service/util/test"
 	"fmt"
 	slog "github.com/go-eden/slf4go"
 	"github.com/go-pg/pg/v10"
@@ -34,7 +34,7 @@ func TestMain(m *testing.M) {
 
 	slog.Info("Running test init func")
 
-	db = test.ConnectDb()
+	db = test2.ConnectDb()
 	repository = NewRepository(db)
 	calendarService = NewService(repository)
